internal/e2e: switch to only one GAE Standard test
